Remove `test_all_sets_modified_image_counts.py` - Remove `all_sets_modified.cfg` too Update `build_workflow.yml` to separate testing steps - Move `complete_run.py` to `tests/integration` - Move `_compare_images()` to `tests/integration/utils.py` - Update `testing.rst` with correct path to `complete_run.py` and `pytest`
