+++ docs/source/dev_guide/testing.rst
@@ -56,7 +56,7 @@ The unit and integration tests are run automatically as part of this.
 Complete run test
 -----------------
 
-``tests/complete_run.py`` checks the images generated by all diagnostics to
+``tests/integration/complete_run.py`` checks the images generated by all diagnostics to
 see if any differ from expected.
 This test is not run as part of the unit test suite, because it relies on a large
 quantity of data found on LCRC (Anvil/Chrysalis).
@@ -78,7 +78,7 @@ Now that you have your changes on LCRC, enter your development environment. Then
     .. code::
 
         pip install . # Install your changes
-        python -m unittest tests/complete_run.py
+        pytest tests/integration/complete_run.py
 
 If this test passes, you're done. If it fails however, that means
 your code has changed what the output looks like.
@@ -108,11 +108,11 @@ then you need to update the expected images.
         find . -type f -name '*.png' > ../image_list_all_sets.txt
         cd ..
 
-Run ``python -m unittest tests/complete_run.py`` again. Now, the test should pass.
+Run ``pytest tests/integration/complete_run.py`` again. Now, the test should pass.
 
 After merging your pull request, edit ``README.md``.
 The version should be the version of E3SM Diags you ran
-``python -m unittest tests/complete_run.py`` with,
-the date should be the date you ran ``python -m unittest tests/complete_run.py`` on,
+``pytest tests/integration/complete_run.py`` with,
+the date should be the date you ran ``pytest tests/integration/complete_run.py`` on,
 and the hash should be for the top commit shown by ``git log`` or on
 https://github.com/E3SM-Project/e3sm_diags/commits/main.
