--- tests/integration/all_sets_modified.cfg
@@ -1,187 +1,0 @@
-[zonal_mean_xy]
-sets = ["zonal_mean_xy"]
-case_id = "ERA-Interim"
-variables = ["T"]
-seasons = ["ANN"]
-plevs = [200.0]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-ref_name = "ERA-Interim"
-reference_name = "ERA-Interim Reanalysis 1979-2015"
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = "ta_ERA-Interim_ANN_198001_201401_climo.nc"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "T_20161118.beta0.FC5COSP.ne30_ne30.edison_ANN_climo.nc"
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-
-
-[zonal_mean_2d]
-sets = ["zonal_mean_2d"]
-case_id = "ERA-Interim"
-variables = ["T"]
-seasons = ["ANN"]
-contour_levels = [180,185,190,200,210,220,230,240,250,260,270,280,290,295,300]
-diff_levels = [-7,-6,-5,-4,-3,-2,-1,1,2,3,4,5,6,7]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-ref_name = "ERA-Interim"
-reference_name = "ERA-Interim Reanalysis 1989-2005"
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = "ta_ERA-Interim_ANN_198001_201401_climo.nc"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "T_20161118.beta0.FC5COSP.ne30_ne30.edison_ANN_climo.nc"
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-plot_log_plevs = False
-plot_plevs = False
-
-
-[meridional_mean_2d]
-sets = ["meridional_mean_2d"]
-case_id = "ERA-Interim"
-variables = ["T"]
-seasons = ["ANN"]
-contour_levels = [180,185,190,200,210,220,230,240,250,260,270,280,290,295,300]
-diff_levels = [-7,-6,-5,-4,-3,-2,-1,1,2,3,4,5,6,7]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-ref_name = "ERA-Interim"
-reference_name = "ERA-Interim Reanalysis 1989-2005"
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = "ta_ERA-Interim_ANN_198001_201401_climo.nc"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "T_20161118.beta0.FC5COSP.ne30_ne30.edison_ANN_climo.nc"
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-plot_log_plevs = False
-plot_plevs = False
-
-
-[lat_lon]
-sets = ["lat_lon"]
-case_id = "ERA-Interim"
-variables = ["T"]
-seasons = ["ANN"]
-plevs = [850.0]
-contour_levels = [240, 245, 250, 255, 260, 265, 270, 275, 280, 285, 290, 295]
-diff_levels = [-10, -7.5, -5, -4, -3, -2, -1, -0.5, 0.5, 1, 2, 3, 4, 5, 7.5, 10]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-ref_name = "ERA-Interim"
-reference_name = "ERA-Interim Reanalysis 1979-2015"
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = "ta_ERA-Interim_ANN_198001_201401_climo.nc"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "T_20161118.beta0.FC5COSP.ne30_ne30.edison_ANN_climo.nc"
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-
-
-[polar]
-sets = ["polar"]
-case_id = "ERA-Interim"
-variables = ["T"]
-seasons = ["ANN"]
-regions = ["polar_S"]
-plevs = [850.0]
-contour_levels = [240, 244, 248, 252, 256, 260, 264, 268, 272]
-diff_levels = [-15, -10, -7.5, -5, -2.5, -1, 1, 2.5, 5, 7.5, 10, 15]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-ref_name = "ERA-Interim"
-reference_name = "ERA-Interim Reanalysis 1979-2015"
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = "ta_ERA-Interim_ANN_198001_201401_climo.nc"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "T_20161118.beta0.FC5COSP.ne30_ne30.edison_ANN_climo.nc"
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-
-
-[cosp_histogram]
-sets = ["cosp_histogram"]
-case_id = "MISR-COSP"
-variables = ["COSP_HISTOGRAM_MISR"]
-seasons = ["ANN"]
-contour_levels = [0,0.5,1.0,1.5,2.0,2.5,3.0,3.5,4.0,4.5,5.0,5.5]
-diff_levels = [-3.0,-2.5,-2.0,-1.5,-1.0,-0.5,0,0.5,1.0,1.5,2.0,2.5,3.0]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-ref_name = "MISRCOSP"
-reference_name = "MISR COSP (2000-2009)"
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = "CLDMISR_ERA-Interim_ANN_198001_201401_climo.nc"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "CLD_MISR_20161118.beta0.FC5COSP.ne30_ne30.edison_ANN_climo.nc"
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-
-
-[area_mean_time_series]
-sets = ["area_mean_time_series"]
-variables = ["TREFHT"]
-
-test_name = "system tests"
-short_test_name = "short_system tests"
-test_data_path = "tests/integration/integration_test_data"
-test_file = "TREFHT_201201_201312.nc"
-start_yr = '2012'
-end_yr = '2013'
-
-backend = "mpl"
-results_dir = "tests/integration/all_sets_results"
-debug = True
-ref_names = []
-ref_timeseries_input = True
-test_timeseries_input = True
-ref_start_yr='2012'
-ref_end_yr='2013'
-test_start_yr='2012'
-test_end_yr='2013'
-
-
-[enso_diags]
-sets = ["enso_diags"]
-case_id = 'TREFHT-response'
-debug = True
-diff_colormap = "BrBG"
-regions = ["20S20N"]
-results_dir = "tests/integration/all_sets_results"
-seasons = ["ANN"]
-start_yr = '2012'
-end_yr = '2013'
-test_data_path = "tests/integration/integration_test_data"
-test_file = 'TREFHT_201201_201312.nc'
-reference_data_path = "tests/integration/integration_test_data"
-ref_file = 'TREFHT_201201_201312.nc'
-test_name = "system tests"
-variables = ["TREFHT"]
-print_statements  = True
-nino_region = "NINO34"
-ref_timeseries_input = True
-test_timeseries_input = True
-ref_start_yr='2012'
-ref_end_yr='2013'
-test_start_yr='2012'
-test_end_yr='2013'
-backend = "mpl"
-plot_type="map"
